Day 16
This solution is a bit dirty and somewhat manual. After getting array of
sets with possible field meanings I noticed a pattern. There is one
field with only one possible value. If we remove this value as an option
for all the other fields, we will get another one like that and so
forth. I decided to simply print sets in convenient order and just go
trough them manually, striking out option until I had all the fields
required to get answer for part 2.